requirements.txt
streamlit>=1.31.0
google-genai>=0.2.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
